chore: remove duplicate file
